+++ projects/ngx-testing-tools/src/lib/components/test-bed/index.ts
@@ -1,2 +1,3 @@
 export * from './models';
+export * from './store';
 export { componentTestBed } from './component-test-bed';
