--- projects/ngx-testing-tools/src/lib/components/test-bed/injected/models/injected-store.model.ts
@@ -1,3 +1,0 @@
-export type InjectionStore<T extends {}> = {
-  injected: T
-}
