+++ projects/ngx-testing-tools/src/lib/components/test-bed/models/component-test-bed.models.ts
@@ -1,10 +1,10 @@
 import { ComponentTestBedFactory } from '../component-test-bed-factory';
-import { InjectionStore } from '../injected/models/injected-store.model';
+import { InjectionStore } from '../store';
 import { ComponentExtraOptions } from './component-extra-options.model';
 import { ComponentTools } from './component-tools.model';
 
-export interface ComponentTestBed<T, I extends InjectionStore<{}> = InjectionStore<{}>> extends ComponentTestBedFn<T, I>, ComponentTestBedFactory<T, I> {}
+export interface ComponentTestBed<T, I extends InjectionStore = InjectionStore> extends ComponentTestBedFn<T, I>, ComponentTestBedFactory<T, I> {}
 
-export type ComponentTestBedFn<T, I extends InjectionStore<{}>> = (assertion: ComponentAssertion<T, I['injected']>, options?: ComponentExtraOptions) => jasmine.ImplementationCallback
+export type ComponentTestBedFn<T, I extends InjectionStore> = (assertion: ComponentAssertion<T, I['injected']>, options?: ComponentExtraOptions) => jasmine.ImplementationCallback
 
 export type ComponentAssertion<T, I extends {}> = (tools: ComponentTools<T, I>, done: DoneFn) => ReturnType<jasmine.ImplementationCallback>
