+++ projects/ngx-testing-tools/src/lib/components/test-bed/store/index.ts
@@ -1,0 +1,1 @@
+export { InjectionStore } from './models/injected-store.model';
