+++ projects/ngx-testing-tools/src/lib/components/test-bed/store/injected.ts
@@ -2,7 +2,7 @@ import { fromInjector } from '../../../injector';
 import { ComponentTestBedFactory } from '../component-test-bed-factory';
 import { InjectionStore } from './models/injected-store.model';
 
-export function buildInjected(factory: ComponentTestBedFactory<unknown>): InjectionStore<{}>['injected'] {
+export function buildInjected(factory: ComponentTestBedFactory<unknown>): InjectionStore['injected'] {
   const injected: InjectionStore<any>['injected'] = {};
   for (const [key, value] of factory['injected'].entries()) {
     injected[value] = fromInjector(factory['fixture'], key);
