+++ projects/ngx-testing-tools/src/lib/components/test-bed/store/models/injected-store.model.ts
@@ -1,0 +1,3 @@
+export type InjectionStore<T extends {} = {}> = {
+  injected: T;
+}
